// src/types/Classes.ts
export enum ClassesEnum {
	Barbarian = 'barbarian',
	Cleric = 'cleric',
	Bard = 'bard',
	Druid = 'druid',
	Monk = 'monk',
	Fighter = 'fighter',
	Paladin = 'paladin',
	Ranger = 'ranger',
	Rogue = 'rogue',
	Sorcerer = 'sorcerer',
	Warlock = 'warlock',
	Wizard = 'wizard'
}

export const Classes = [
	{ index: ClassesEnum.Barbarian, name: 'Barbarian' },
	{ index: ClassesEnum.Cleric, name: 'Cleric' },
	{ index: ClassesEnum.Bard, name: 'Bard' },
	{ index: ClassesEnum.Druid, name: 'Druid' },
	{ index: ClassesEnum.Monk, name: 'Monk' },
	{ index: ClassesEnum.Fighter, name: 'Fighter' },
	{ index: ClassesEnum.Paladin, name: 'Paladin' },
	{ index: ClassesEnum.Ranger, name: 'Ranger' },
	{ index: ClassesEnum.Rogue, name: 'Rogue' },
	{ index: ClassesEnum.Sorcerer, name: 'Sorcerer' },
	{ index: ClassesEnum.Warlock, name: 'Warlock' },
	{ index: ClassesEnum.Wizard, name: 'Wizard' }
];